Add ability to generate tools repositories on Fedora.
Added third section in repository creation script
which activates on Fedora. This is tested on Fedora-16
and Fedora-17. Note that these have yum on different paths:
F-16: /usr/bin/yum
F-17: /bin/yum
The repo lines are all generated, because there is no
easy way of adding repo via command, as yum-config-manager
can only add repo and not specify it's name, which is
then needed for later add-options commands, so it would
become complex to do it all via commands.
I also replaced executable file tests to be -x, to make
the matching case more exact, and added title comment
in the code blocks.

Change-Id: I2d03bfdc41d93b8223b10d467bddb57935815fcd