Increase OBS build monitoring timeout
For big projects with a lot of target repositories default timeout 60
seconds is not enough. Sometimes it takes more than one minute for OBS
to start a build. This change is increasing timeout between the time
when Jenkins job starts and the time when it starts checking OBS build
status.

Change-Id: I796d13fdc222e5332bf0da268853d0835cfcd281