Reverted: Running pylint in top source dir again
As we agreed with Markus I'm reverting previous change. run_tests now
will run pylint in the top source directory as it was before. It caused
pylint false positives and we decided to workaround them by running
pylint in the same directory as .py file under test. It looks like it
was not the best decision. Now it again causes problems. Let's test if
we can fix original issues properly with Markus help. This change should
only be deployed in production after careful testing with all production
projects, which use pylint.

Change-Id: I3539deabde66b3c510d8ad7645df6285bbb7f60e
Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>