Reimplemented waiting for build results
Now build-packages script uses more intelligent way to wait for the
build results. First it waits until project and package exists in OBS.
After that it waits until build results are available for the package
and only after that analyses them.

It's quite big change, so it requires serious pre-production testing.

Change-Id: I6c118674fdd452ccf1aebe1262fa8aa00dc5ec01
Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>