Used Gerrit change as a part of OBS project name
OBS project is created for each gerrit change and patchset. Currently
Jenkins build number is used as a suffix in the project name. However,
It would be much more convenient for developers to see Gerrit change
and patchset number as a part of OBS project name instead.
Gerrit change is a reason for the build and it's more intuitive and
understandable by developers than Jenkins build number.

Note that In previous implementation jenkins build was always uniq, but
now project name for restarted build will be the same as for original
build as they both base on the same Gerrit change. This is taken care of
by adding checks for existence of target project and package in build-package
script. Project and package will be created only if they don't exist
yet.

Change-Id: Ide1656075963044a9bfd5edc803b4dbc747624c7
Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>