Add ABORT signal handling in Cleanup handler
When Jenkins job is aborted from GUI, this script receives
an ABRT signal, and this was not covered by trap handler,
caused leftover of the mounted filesystem which kvm used.
With this change we add ABRT to covered signals by trap handler.
But as qemu-kvm process remains running and keeps mount busy,
we also add lazy umount option. This makes it possible
to umount while qemu-kvm is still running (it exits later
and kernel handles the waiting umount request).
To reduce sudo rules, we use similar umount in all places,
lazy option does not harm for usual umount either.

Change-Id: Ic3694435c6e296583d8d4f94b0326d2d0441a4e4