Disable error checking for building in OBS
With separated Builder and Tester roles,
if build fails, none of the Tester runs, and the plan
was to always force success of Builder with exit 0,
but this does not do what we need, since the shell is
running with set -e and in case of failure we never reach
this exit stmt, shell exits with failure before it.
To make it work as we planned, we need set +e before build call.

Change-Id: I229cf2c896d743e242903e8d184f50b77411c295