Add options to mount commands about ext4 type and verbosity
There was possibility to race condition where blkid was not ready
in time and mount failed to get file system type from blkid.
As ext4 also happens to be not listed in /etc/filesystems,
mount cmd failed to mount HDB image in rare cases.
Giving -t ext4 to mount should avoid such errors.
We also give -v option to mount in hope to see better
information in case of problems in the future.

Change-Id: I8ab4d57517327c4c8c713ebf32135f4822e9ea5e