Workarounded Gerrit Trigger behaviour
For 'Ref Updated' events Gerrit Trigger plugin doesn't set
GERRIT_REFSPEC and GERRIT_BRANCH variables. Because of that it's not
possible to setup Gerrit Trigger to work with Git plugin if job
processes 'Ref Updated' event.

This workaround resets git tree to GERRIT_REFSPEC or GERRIT_REFNAME
depending on the event allowing not to use GERRIT_REFSPEC and
GERRIT_BRANCH in job configuration. It also replaces GERRIT_BRANCH to
GERRIT_REFNAME in the code, which processes 'Ref Updated' events.

Change-Id: Iee7fd9b75736a96da7577a0a6ff0f6d03249d910
Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>