In case of failed debug job, disable autorun and keep images
It is convenient to have the failed setup stored,
but we were destroying everything always by unmounting
ramfs. Lets keep the images for post-analysis.
To avoid autorun and shutdown kicking in, we rename
run script in HDB image. For easier KVM session start,
we generate run script. Note, we need to deal
with cumulating images on workers, rotate and delete.

Change-Id: Ia070ac7c8ef878981727b290a6ad3467393e44a2