Generate git-buildpackage source service
If git url (optional parameter -u) is provided to the script it
generates _service file for git-buildpackage source service. Otherwise
old approach is used: run make all in packaging/.

This way we can use different approaches for different packages and move
gradually to using source service as a preferred solution.

Change-Id: I69e556927eac201d5acc0a89cbf052b813f56e1b
Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>