Implemented handling of test-<target branch>
When commit is pushed to test-devel, test-master or test-release branch
the script will behave the same way as for 'patchset created' events:
Linked project home:testerTools-<package>-test-<devel|master|release>
will be created from correspondent OBS base project:
    for test-devel branch: <main project>:Devel
    for test-release branch: <main project>:Pre-release and
    for test-master branch: <main project>
and sources will be uploaded to that project for the build.

This feature allows developer to trigger Tester job when pushing for
review is not feasible, for example when dealing with lot of changes
from upstream. Another reason of using this is to be able to test
changes without bothering people by review requests from Gerrit.

Change-Id: I788504049f95165b60be8b38fba883a5c0e66363