KVM session uses HDA in snapshot mode
Instead of copying HDA image to 6G ram disk,
we use snapshot mode for HDA. All sessions of
same target tester share same HDA seed image and no changes
are written there, changes are lost. This makes it
possible to run more workers on one server and reduces
RAM usage as big ram disk is not created.
Because of this, we can not store FAILED debug-mode images
any more for later libvirt machine creation,
removing that functionality, it was not used anyway.

Change-Id: I0dec37ceea31034b8316c77336b0b9cea0225fb9