run_tests: disable pylint's I0011
Every time project maintainers disable a pylint
recommendation/warning/error/etc, pylint still produces something like:

I: 25,0: Locally disabling R0904

and this increases the pylint complaint scores. So the result is that it is
impossible to completely ignore a pylint recommendation, because the score will
be increased anyway.

The effect of this is that as the project grows, the pylint complaint score
grows too. This is quite annoying for maintainers, and our system exists to
help maintainers, not annoy them.

For example, I cannot make my pylint complaints score to be 0 in my project,
because I have a bunch of this 'Locally disabling' messages which keep the
score non-zero and constantly increasing. And the effect is that I start
treating the checking system as "unfriendly" and simply do not look at the
check result, because I know there will be a bunch of uninteresting messages.

Let's make it possible for maintainers to keep their projects clean. If the
maintainer believe a pylint recommendation is a false-positive, lets allow
him/her completely disabling it.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>