Start qemu monitor at unix socket instead of VNC
Starting VNC with static number which equal to Jenkins slot number
created possibility of collision in erroneous case when
previous qemu session had not exited.
This change starts qemu-kvm using "-vnc none" which sets vnc
to standby mode and can be activated from qemu monitor.
So we have to start monitor from the start, which in turn
listend on unix socket. This is not tied to any global resource
but is a file system object under isolated part of the session,
therefore eliminating risk of collision.

Change-Id: I4f8a503a08d8d1e59b96716846b17e781db89c6e