run-test-kvm: Make it possible to define "test requires" in env
Parse extra package dependencies (required by unit tests) from
TEST_REQUIRES environment variable if it is defined. This makes it
possible to define the extra dependencies in the jenkins job instead of
polluting the git repository with internal testing data.

The TEST_REQUIRES environment variable must have the same syntax as the
.test-requires file would have.

Change-Id: If88b77e8cf21f4f7d7d5a1fd3f8a4b20922de7dd
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>