build-package: copy package build meta from source pkg
Copy the package build flags (from the package meta information) from
the base package to the target package after branching a package in OBS.
This way we only build for the same targets as in the base project and
do not get false build failures.

Newly created packages are not affected, of course, because there is no
"orig meta" from where to copy.

Change-Id: I04bbcf486eb9dddff94f1e73690a6aeb9dd94970
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>