Tester VM session output comes to stdout and directly to Jenkins
runtester output is written out as serial console output
while tester runs, and comes to Jenkins log in real time,
making the session more user-friendly to monitor and troubleshoot.
In tester VM images, runtester needs to be updated to version
given here.
crontab line needs to be:
@reboot /root/runtester > /dev/ttyS0 2>&1
Note: kernel flags need not to be changed!
Thanks Junchun Guan for giving main ideas for this.

Change-Id: I50d9479280aca577530439beee58c8ed2d27cc62