Add base repo lists as sub-packages
We move to model where tester VMs do not have any repositories
configured. Base and update repository config is inserted
during VM preparation phase. The repo config values are
defined in the file base-repos-XYZ residing in /etc on tester host,
which are copied to VM HDB before starting VM.
This file can be tailored to point to mirrors in different envs.

Change-Id: I2d76928d7670dfec3d4c41e56771b195952453b7