numactl binary was mentioned twice
compose_numa_opt function uses echo to create return value,
but uses also 'which' to get the binary name.
Output of "which" becomes first part of function return value.
But we used to also mention numactl in explicit return string.
As result, numactl was present twice in returned string:
/usr/bin/numactl numactl --options
As numactl does not fail with unknown option, this did not
cause failures. Lets still clean this.

Change-Id: I4d6f3e582b95908fb1a2f85751b92239a69e614e